requirements.txt
fastapi>=0.109.0
uvicorn>=0.25.0
python-dotenv>=1.0.0
azure-cognitiveservices-speech>=1.31.0
httpx>=0.25.2
openai>=1.11.0
async-timeout>=4.0.3
websockets>=11.0.3 